Reorganize API classes.
