+++ src/API/NoteActions.php
@@ -7,7 +7,7 @@
  * @package WooCommerce Admin/API
  */
 
-namespace Automattic\WooCommerce\Admin\API\Notes\Actions;
+namespace Automattic\WooCommerce\Admin\API;
 
 defined( 'ABSPATH' ) || exit;
 
@@ -19,7 +19,7 @@ use \Automattic\WooCommerce\Admin\Notes\WC_Admin_Notes;
  * @package WooCommerce/API
  * @extends WC_REST_CRUD_Controller
  */
-class Controller extends \Automattic\WooCommerce\Admin\API\Notes\Controller {
+class NoteActions extends Notes {
 
 	/**
 	 * Register the routes for admin notes.
